fix: consolidate Models package files for pkg-file-count CI gate
Merge characterization tests and small contract slices, and combine internal/service test helpers so pkg/services/models and internal/service stay within the 15-file package limit.
