+++ pkg/services/models/local_execution_contract.go
@@ -1,6 +1,11 @@
 package models
 
-import "github.com/portpowered/infinite-you/pkg/services/work"
+import (
+	"fmt"
+	"strings"
+
+	"github.com/portpowered/infinite-you/pkg/services/work"
+)
 
 // LocalInvocationRequest is the plain infer request on the Models root.
 // Peers supply Worker/dispatch/bindings vocabulary without importing nested
@@ -52,4 +57,15 @@ type LocalResource struct {
 type LocalInvocationResult struct {
 	Handled bool
 	Content string
+}
+
+// ValidateLocalInvocationRequest checks the plain infer/local-invocation
+// request. Managed-runtime workers with an empty Model fail closed as
+// ErrNotFound without touching nested inference or local-execution packages.
+// Non-managed workers remain valid so InvokeLocal can return Handled=false.
+func ValidateLocalInvocationRequest(request LocalInvocationRequest) error {
+	if request.Worker.UsesManagedRuntime() && strings.TrimSpace(request.Worker.Model) == "" {
+		return fmt.Errorf("%w: empty managed runtime model name", ErrNotFound)
+	}
+	return nil
 }
